pyproject.toml
[tool.poetry]
name = "instagram-app"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.11.11"
aiohttp-socks = "^0.10.1"
aiograpi = "^0.0.3"
requests = {extras = ["socks"], version = "^2.32.3"}
configparser = "^7.1.0"
fake-useragent = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
